requirements.txt
Pillow>=9.0.0
imageio>=2.30.0
imageio-ffmpeg>=0.5.0
